requirements.txt
asgiref==3.3.4
chardet==4.0.0
Django==3.2.3
django-bootstrap-icons==0.6.0
django-bs-icons==0.2.2
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
future==0.18.2
idna==2.10
pefile==2021.5.13
phonenumbers==8.12.23
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
